requirements.txt
aiosqlite==0.17.0
anyio==3.3.2
APScheduler==3.8.1
arrow==1.2.0
asgiref==3.4.1
asyncio==3.4.3
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.68.2
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.18.2
hyperframe==5.2.0
idna==3.2
Jinja2==3.0.2
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
nb-cli==0.5.0
nonebot-adapter-cqhttp==2.0.0a16
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a16
poyo==0.5.0
prompt-toolkit==3.0.20
pydantic==1.8.2
pyfiglet==0.8.post1
pygtrie==2.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
spark-parser==1.8.9
SQLAlchemy==1.4.41
starlette==0.14.2
text-unidecode==1.3
tomlkit==0.7.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
uncompyle==2.0.0
uncompyle6==3.7.4
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
win32-setctime==1.0.3
xdis==5.0.13
